Add support for background processing using sidekiq & active job
The model worker will immediately run in the background after a new job
is created. It will send the 3D model file to the hub and do other computations as needed.
There is a model_processing flag now added to the job model which is set to TRUE while the
model worker is has not yet completed, and FALSE when it has completed.